doTogether/requirements.txt
backcall==0.1.0
certifi==2018.8.13
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
decorator==4.3.0
Django==2.1
django-extensions==2.1.0
django-oauth-toolkit==1.2.0
django-rest-swagger==2.2.0
djangorestframework==3.8.2
djangorestframework-simplejwt==3.2.3
idna==2.7
ipython==6.5.0
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.12.1
Jinja2==2.10
MarkupSafe==1.0
openapi-codec==1.3.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.2.0
prompt-toolkit==1.0.15
psycopg2-binary==2.7.5
ptyprocess==0.6.0
Pygments==2.2.0
pytz==2018.5
requests==2.19.1
simplegeneric==0.8.1
simplejson==3.16.0
six==1.11.0
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.23
wcwidth==0.1.7
